Move redis test client, update cache field
